# movies_project/async_api/openapi.json
{"openapi":"3.1.0","info":{"title":"movies","version":"0.1.0"},"paths":{"/api/v1/films/search":{"get":{"summary":"Поиск по фильмам.","description":"Полнотекстовый поиск фильма по указанному параметру query в запросе и по переданным uuid жанра в параметре genre и персоны в параметре person, с сортировкой по параметру sort и пагинацией.","operationId":"film_search_api_v1_films_search_get","parameters":[{"name":"query","in":"query","required":false,"schema":{"type":"string","description":"Текст для поиска","title":"Query"},"description":"Текст для поиска"},{"name":"genre","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"description":"Жанр","title":"Genre"},"description":"Жанр"},{"name":"person","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"description":"Персона","title":"Person"},"description":"Персона"},{"name":"page_number","in":"query","required":false,"schema":{"type":"integer","maximum":1000,"minimum":1,"description":"Номер страницы","default":1,"title":"Page Number"},"description":"Номер страницы"},{"name":"page_size","in":"query","required":false,"schema":{"type":"integer","maximum":50,"minimum":1,"description":"Размер страницы","default":10,"title":"Page Size"},"description":"Размер страницы"},{"name":"sort","in":"query","required":false,"schema":{"type":"string","description":"Сортировка по убыванию -id по возрастанию id","default":"id","title":"Sort"},"description":"Сортировка по убыванию -id по возрастанию id"}],"responses":{"200":{"description":"Список фильмов.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FilmInfoResponse"},"title":"Response Film Search Api V1 Films Search Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/films/{id_uuid}":{"get":{"summary":"Информация о фильме.","description":"Детальная информация о фильме и список жанров, актёров, режиссёров и писателей.","operationId":"film_id_api_v1_films__id_uuid__get","parameters":[{"name":"id_uuid","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Id Uuid"}}],"responses":{"200":{"description":"Фильм с информацией по всем имеющимися полям.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FilmFullInfoResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/films":{"get":{"summary":"Главная страница фильмов.","description":"На ней выводятся популярные фильмы, с возможностью фильтрации по жанру.","operationId":"film_list_api_v1_films_get","parameters":[{"name":"genre","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"description":"UUID жанра","title":"Genre"},"description":"UUID жанра"},{"name":"page_number","in":"query","required":false,"schema":{"type":"integer","maximum":1000,"minimum":1,"description":"Номер страницы","default":1,"title":"Page Number"},"description":"Номер страницы"},{"name":"page_size","in":"query","required":false,"schema":{"type":"integer","maximum":50,"minimum":1,"description":"Размер страницы","default":10,"title":"Page Size"},"description":"Размер страницы"},{"name":"sort","in":"query","required":false,"schema":{"type":"string","description":"Сортировка по убыванию -id по возрастанию id","default":"id","title":"Sort"},"description":"Сортировка по убыванию -id по возрастанию id"}],"responses":{"200":{"description":"Список фильмов.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FilmInfoResponse"},"title":"Response Film List Api V1 Films Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/genres/{id_uuid}":{"get":{"summary":"Информация о жанре.","operationId":"genre_details_api_v1_genres__id_uuid__get","parameters":[{"name":"id_uuid","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Id Uuid"}}],"responses":{"200":{"description":"Информация о жанре.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenreInfoResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/genres":{"get":{"summary":"Список жанров.","operationId":"genres_details_list_api_v1_genres_get","parameters":[{"name":"page_number","in":"query","required":false,"schema":{"type":"integer","maximum":1000,"minimum":1,"description":"Номер страницы","default":1,"title":"Page Number"},"description":"Номер страницы"},{"name":"page_size","in":"query","required":false,"schema":{"type":"integer","maximum":50,"minimum":1,"description":"Размер страницы","default":10,"title":"Page Size"},"description":"Размер страницы"},{"name":"sort","in":"query","required":false,"schema":{"type":"string","description":"Сортировка по убыванию -id по возрастанию id","default":"id","title":"Sort"},"description":"Сортировка по убыванию -id по возрастанию id"}],"responses":{"200":{"description":"Список жанров.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GenreInfoResponse"},"title":"Response Genres Details List Api V1 Genres Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/persons/{id_uuid}/film":{"get":{"summary":"Фильмы по персоне.","description":"Получение списка фильмов, в которых участвовала данная персона.","operationId":"persons_film_api_v1_persons__id_uuid__film_get","parameters":[{"name":"id_uuid","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Id Uuid"}},{"name":"page_number","in":"query","required":false,"schema":{"type":"integer","maximum":1000,"minimum":1,"description":"Номер страницы","default":1,"title":"Page Number"},"description":"Номер страницы"},{"name":"page_size","in":"query","required":false,"schema":{"type":"integer","maximum":50,"minimum":1,"description":"Размер страницы","default":10,"title":"Page Size"},"description":"Размер страницы"},{"name":"sort","in":"query","required":false,"schema":{"type":"string","description":"Сортировка по убыванию -id по возрастанию id","default":"id","title":"Sort"},"description":"Сортировка по убыванию -id по возрастанию id"}],"responses":{"200":{"description":"Список фильмов.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PersonFilmResponse"},"title":"Response Persons Film Api V1 Persons  Id Uuid  Film Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/persons/search":{"get":{"summary":"Поиск по персонам.","description":"Полнотекстовый поиск персоны по указанному параметру query в запросе.","operationId":"persons_list_api_v1_persons_search_get","parameters":[{"name":"query","in":"query","required":false,"schema":{"type":"string","description":"Запрос","title":"Query"},"description":"Запрос"},{"name":"page_number","in":"query","required":false,"schema":{"type":"integer","maximum":1000,"minimum":1,"description":"Номер страницы","default":1,"title":"Page Number"},"description":"Номер страницы"},{"name":"page_size","in":"query","required":false,"schema":{"type":"integer","maximum":50,"minimum":1,"description":"Размер страницы","default":10,"title":"Page Size"},"description":"Размер страницы"},{"name":"sort","in":"query","required":false,"schema":{"type":"string","description":"Сортировка по убыванию -id по возрастанию id","default":"id","title":"Sort"},"description":"Сортировка по убыванию -id по возрастанию id"}],"responses":{"200":{"description":"Список персон.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PersonInfoResponse"},"title":"Response Persons List Api V1 Persons Search Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/persons/{id_uuid}":{"get":{"summary":"Информация о персоне.","description":"Получение информации по персоне по её uuid.","operationId":"persons_api_v1_persons__id_uuid__get","parameters":[{"name":"id_uuid","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Id Uuid"}}],"responses":{"200":{"description":"Детальная информация о персоне и список фильмов со списком ролей в них.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PersonInfoResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"FilmFullInfoResponse":{"properties":{"uuid":{"type":"string","title":"Uuid"},"title":{"type":"string","title":"Title"},"imdb_rating":{"type":"number","title":"Imdb Rating"},"description":{"type":"string","title":"Description"},"genres":{"items":{"$ref":"#/components/schemas/Name"},"type":"array","title":"Genres"},"actors":{"items":{"$ref":"#/components/schemas/FullName"},"type":"array","title":"Actors"},"writers":{"items":{"$ref":"#/components/schemas/FullName"},"type":"array","title":"Writers"},"directors":{"items":{"$ref":"#/components/schemas/FullName"},"type":"array","title":"Directors"}},"type":"object","required":["uuid","title","imdb_rating","description","genres","actors","writers","directors"],"title":"FilmFullInfoResponse"},"FilmInfoResponse":{"properties":{"uuid":{"type":"string","title":"Uuid"},"title":{"type":"string","title":"Title"},"imdb_rating":{"type":"number","title":"Imdb Rating"}},"type":"object","required":["uuid","title","imdb_rating"],"title":"FilmInfoResponse"},"Films":{"properties":{"uuid":{"type":"string","title":"Uuid"},"roles":{"items":{"type":"string"},"type":"array","title":"Roles"}},"type":"object","required":["uuid","roles"],"title":"Films"},"FullName":{"properties":{"uuid":{"type":"string","title":"Uuid"},"full_name":{"type":"string","title":"Full Name"}},"type":"object","required":["uuid","full_name"],"title":"FullName"},"GenreInfoResponse":{"properties":{"uuid":{"type":"string","title":"Uuid"},"name":{"type":"string","title":"Name"}},"type":"object","required":["uuid","name"],"title":"GenreInfoResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Name":{"properties":{"uuid":{"type":"string","title":"Uuid"},"name":{"type":"string","title":"Name"}},"type":"object","required":["uuid","name"],"title":"Name"},"PersonFilmResponse":{"properties":{"uuid":{"type":"string","title":"Uuid"},"title":{"type":"string","title":"Title"},"imdb_rating":{"type":"number","title":"Imdb Rating"}},"type":"object","required":["uuid","title","imdb_rating"],"title":"PersonFilmResponse"},"PersonInfoResponse":{"properties":{"uuid":{"type":"string","title":"Uuid"},"full_name":{"type":"string","title":"Full Name"},"films":{"items":{"$ref":"#/components/schemas/Films"},"type":"array","title":"Films"}},"type":"object","required":["uuid","full_name","films"],"title":"PersonInfoResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}