# movies_project/async_api/src/models/response_film.py
from models.base import FullName, Name, FilmInfo


class FilmFullInfoResponse(FilmInfo):
    description: str
    genres: list[Name]
    actors: list[FullName]
    writers: list[FullName]
    directors: list[FullName]